Fix: Page 확장자 통일
